requirements.txt
chardet==4.0.0
charset-normalizer==2.0.12
PyYAML==6.0
